+++ appdaemon/plugins/mqtt/mqttplugin.py
@@ -278,7 +278,7 @@ class MqttPlugin:
                     self.AD.log("INFO", "{}: MQTT Plugin initialization complete".format(self.name))
                 else:
                     if not already_notified and already_initialized:
-                        self.AD.notify_plugin_stopped(self.name)
+                        self.AD.notify_plugin_stopped(self.name, self.namespace)
                         self.AD.log("CRITICAL", "{}: MQTT Plugin Stopped Unexpectedly".format(self.name))
                         already_notified = True
                         already_initialized = False